+++ customizer.php
@@ -5,6 +5,11 @@ function awada_customizer_preview_js()
     wp_enqueue_script('custom_css_preview', get_template_directory_uri() . '/js/customize-preview.js', array('customize-preview', 'jquery'));
 }
 add_action('customize_preview_init', 'awada_customizer_preview_js');
+function awada_controls_enqueue_scripts($wp_customize)
+{
+    wp_enqueue_script('awada_customize_control', get_template_directory_uri() . '/js/customize_custom_control.js', array('jquery', 'customize-controls'));
+}
+add_action('customize_controls_enqueue_scripts', 'awada_controls_enqueue_scripts', 11);
 /* Modifying existing controls */
 add_action( 'customize_register', function( $wp_customize ) {
     $wp_customize->get_setting( 'blogname' )->transport          = 'postMessage';
@@ -34,55 +39,9 @@ function awada_render_footer_copyright(){
 		<p><span id="copyright_text">'.esc_attr($awada_theme_options['footer_copyright']).'</span><span id="developed_by_text"> '.esc_attr($awada_theme_options['developed_by_text']).' </span><a id="copyright_link" href="'. esc_url($awada_theme_options['developed_by_link']).'"><span id="copyright_link_text">'.esc_attr($awada_theme_options['developed_by_link_text']).'</span></a></p>
 	</div>';
 }
-function awada_partial_refresh_callout_title(){
-	$awada_theme_options = awada_theme_options();
-	echo '<h2 id="callout_title">'.$awada_theme_options['home_callout_title'].'</h2>';
-}
-function awada_partial_refresh_callout_desc(){
-	$awada_theme_options = awada_theme_options();
-	echo '<p id="callout_description" class="lead">'.$awada_theme_options['home_callout_description'].'</p>';
-}
 function awada_partial_refresh_callout_btn(){
 	$awada_theme_options = awada_theme_options();
 	echo '<a id="callout_link" class="btn btn-dark btn-lg btn-shadow" href="'.$awada_theme_options['callout_btn_link'].'" target="_blank"><span class="intro_text">'.$awada_theme_options['callout_btn_text'].'</span></a>';
-}
-function awada_partial_refresh_blog_title(){
-	$awada_theme_options = awada_theme_options();
-	echo '<h2 id="blog_heading">'.$awada_theme_options['home_blog_title'].'</h2>';
-}
-function awada_partial_refresh_blog_desc(){
-	$awada_theme_options = awada_theme_options();
-	echo '<p id="blog_description">'.$awada_theme_options['home_blog_description'].'</p>';
-}
-function awada_partial_refresh_service_head(){
-	$awada_theme_options = awada_theme_options();
-	echo '<h2 id="home_service_title">'.$awada_theme_options['home_service_title'].'</h2>';
-}
-function awada_partial_refresh_service_desc(){
-	$awada_theme_options = awada_theme_options();
-	echo '<p id="home_service_description">'.$awada_theme_options['home_service_description'].'</p>';
-}
-function awada_partial_refresh_service_icon($i){
-	$act='';
-	if($i==2){ $act = 'active'; }
-	$awada_theme_options = awada_theme_options();
-	$out = '<div id="service_box_'.$i.'" class="service-icon-square  '.$act.'">';
-	$out.= '<i id="service_icon_'.esc_attr($i).'" class="'.esc_attr($awada_theme_options['service_icon_'.$i]).'"></i>';
-	$out.='</div>';
-	echo $out;
-}
-function awada_partial_refresh_service_title($i){
-	$awada_theme_options = awada_theme_options();
-	echo $out = '<h3 id="service_title_'.esc_attr($i).'" >'.esc_attr($awada_theme_options['service_title_'.$i]).'</h3>';
-}
-function awada_partial_refresh_service_text($i){
-	$awada_theme_options = awada_theme_options();
-	echo $out = '<p id="service_description_'.esc_attr($i).'" >'.esc_attr($awada_theme_options['service_text_'.$i]).'</p>';
-	
-}
-function awada_partial_refresh_service_link($i){
-	$awada_theme_options = awada_theme_options();
-	echo $out = '<a id="service_link_'.esc_attr($i).'" class="readmore" href="'.esc_attr($awada_theme_options['service_link_'.$i]).'">'.__('Read More', 'awada').'</a>';
 }
 function awada_render_contact_email(){
 	$awada_theme_options = awada_theme_options();
@@ -94,28 +53,7 @@ function awada_render_contact_phone(){
 	$out = '<i class="fa fa-phone-square"></i> '.$awada_theme_options['contact_phone'];
 	return $out;
 }
-if(!function_exists('awada_get_post_select')):
-	function awada_get_post_select() {
-	$all_posts = wp_count_posts('post')->publish;
-	$latest = new WP_Query( array(
-	'post_type'   => 'post',
-	'post_per_page'=>$all_posts,
-	'post_status' => 'publish',
-	'orderby'     => 'date',
-	'order'       => 'DESC'
-	));
-	  $results;
-	  if(!empty($latest)):
-		  $results['default'] = __('Select Post','awada');
-		  while( $latest->have_posts() ) { $latest->the_post();
-			$results[get_the_id()] = get_the_title();
-			
-		  }
-	  endif;
-	  
-	  return $results;
-	}
-endif;
+
 /* Add Customizer Panel */
 $awada_theme_options = awada_theme_options();
 Kirki::add_config('awada_theme', array(
@@ -222,7 +160,6 @@ Kirki::add_field( 'awada_theme', array(
     'section'     => 'general_sec',
     'default'     => 'green.css',
     'priority'    => 10,
-    'multiple'    => 3,
     'choices'     => array(
         'default.css' => array(
             'label'    => __('Default','awada'),
@@ -271,22 +208,6 @@ Kirki::add_field('awada_theme', array(
         ),
     ),
     'sanitize_callback' => 'awada_sanitize_number',
-));
-
-Kirki::add_field('awada_theme', array(
-    'settings' => 'custom_css',
-    'label' => __('Custom Css Editor', 'awada'),
-    'section' => 'general_sec',
-    'type' => 'code',
-    'priority' => 10,
-    'default' => '',
-    'sanitize_callback'    => 'wp_filter_nohtml_kses',
-    'sanitize_js_callback' => 'wp_filter_nohtml_kses',
-	'choices'     => array(
-		'language' => 'css',
-		'theme'    => 'monokai',
-		'label'=>'Open Css Editor'
-	),
 ));
 
 /* Typography */
@@ -540,34 +461,123 @@ Kirki::add_field('awada_theme', array(
 	'transport'         => 'postMessage',
     'sanitize_callback' => 'awada_sanitize_checkbox',
 ));
-$args = array(
-    'posts_per_page'   => -1,
-    'offset'           => 0,
-    'orderby'          => 'date',
-    'order'            => 'DESC',
-    'post_type'        => 'post',
-    'post_status'      => 'publish',
-    'post__not_in'     => get_option('sticky_posts'),
-    'suppress_filters' => true,
-);
-$posts_list = get_posts($args);
-foreach ($posts_list as $post) {
-    $posts[$post->ID] = $post->post_title;
-}
-Kirki::add_field('awada_theme', array(
-    'type'              => 'select',
-    'settings'          => 'home_slider_posts',
-    'label'             => __('Select Posts to be Shown in Slider', 'awada'),
-    'help'              => __('You can also be able to drag-n-drop the selected options and rearrange their order for maximum flexibility', 'awada'),
-    'section'           => 'slider_sec',
-    'priority'          => 10,
-    'default'           => 0,
-    'choices'           => $posts,
-    'multiple'          => 4,
-    'sanitize_callback' => 'awada_sanitize_selected',
-	
-));
 
+$num = array(1 => 'One', 2 => 'Two', 3 => 'Three');
+for ($i = 1; $i <= 3; $i++) {
+    Kirki::add_field('awada_theme', array(
+        'settings'          => 'slider_heading_' . $i,
+        'label'             => sprintf(__('Slider %s', 'awada'), $num[$i]),
+        'section'           => 'slider_sec',
+        'type'              => 'heading',
+        'priority'          => 10,
+        'sanitize_callback' => 'awada_sanitize_text',
+    ));
+    Kirki::add_field('awada_theme', array(
+        'settings'    => 'slider_img_' . $i,
+        'label'       => sprintf(__('Slider %s Image', 'awada'), $num[$i]),
+        'description' => __('Recommended image size is 1600x500px', 'awada'),
+        'type'        => 'image',
+        'section'     => 'slider_sec',
+        'priority'    => 10,
+        'default'     => get_template_directory_uri() . '/images/slider/s'.$i.'.jpg',
+    ));
+    Kirki::add_field('awada_theme', array(
+        'settings'          => 'slider_title_' . $i,
+        'label'             => sprintf(__('Slider %s Title', 'awada'), $num[$i]),
+        'type'              => 'text',
+        'default'           => $awada_theme_options['slider_title_' . $i],
+        'section'           => 'slider_sec',
+        'priority'          => 10,
+        'sanitize_callback' => 'awada_sanitize_text',
+    ));
+    Kirki::add_field('awada_theme', array(
+        'settings'          => 'slider_subtitle_' . $i,
+        'label'             => sprintf(__('Slider %s Subitle', 'awada'), $num[$i]),
+        'type'              => 'text',
+        'default'           => $awada_theme_options['slider_subtitle_' . $i],
+        'section'           => 'slider_sec',
+        'priority'          => 10,
+        'sanitize_callback' => 'awada_sanitize_text',
+    ));
+    Kirki::add_field('awada_theme', array(
+        'type'     => 'radio-buttonset',
+        'settings' => 'slider_post_page_' . $i,
+        'label'    => __('Link this slide to a post or page', 'awada'),
+        'section'  => 'slider_sec',
+        'default'  => 'post',
+        'priority' => 10,
+        'choices'  => array(
+            'post' => esc_attr__('Post', 'awada'),
+            'page' => esc_attr__('Page', 'awada'),
+        ),
+    ));
+
+    Kirki::add_field('awada_theme', array(
+        'type'            => 'select',
+        'settings'        => 'post_slider_' . $i,
+        'label'           => __('Select a post', 'awada'),
+        'section'         => 'slider_sec',
+        'priority'        => 10,
+        'choices'         => Kirki_Helper::get_posts(array('posts_per_page' => -1, 'orderby' => 'date', 'order' => 'DESC', 'post_type' => 'post', 'post_status' => 'publish')),
+        'active_callback' => array(
+            array(
+                'setting'  => 'slider_post_page_' . $i,
+                'operator' => '==',
+                'value'    => 'post',
+            ),
+        ),
+    ));
+    Kirki::add_field('awada_theme', array(
+        'settings'          => 'page_slider_' . $i,
+        'label'             => __('Select a page', 'awada'),
+        'type'              => 'dropdown-pages',
+        'section'           => 'slider_sec',
+        'priority'          => 10,
+        'sanitize_callback' => 'awada_sanitize_number',
+        'active_callback'   => array(
+            array(
+                'setting'  => 'slider_post_page_' . $i,
+                'operator' => '==',
+                'value'    => 'page',
+            ),
+        ),
+    ));
+    Kirki::add_field('awada_theme', array(
+        'settings'          => 'slider_readmore_' . $i,
+        'label'             => __('Read more button text', 'awada'),
+        'type'              => 'text',
+        'section'           => 'slider_sec',
+        'priority'          => 10,
+        'default'           => __('Read More', 'awada'),
+        'sanitize_callback' => 'awada_sanitize_text',
+        'active_callback'   => function () use ($i) {
+            $awada_options = get_option('awada_theme_options');
+            if (isset($awada_options['slider_post_page_' . $i]) && $awada_options['slider_post_page_' . $i] == 'post') {
+                if (isset($awada_options['post_slider_' . $i]) && $awada_options['post_slider_' . $i] != 0) {return true;}
+            } else {
+                if (isset($awada_options['page_slider_' . $i]) && $awada_options['page_slider_' . $i] != 0) {return true;}
+            }
+            return false;
+        },
+    ));
+}
+
+Kirki::add_field('awada_theme', array(
+    'settings'          => 'slider_plugin_heading',
+    'label'             => __('Or Put Slider Plugin Shortcode here', 'awada'),
+    'section'           => 'slider_sec',
+    'type'              => 'heading',
+    'priority'          => 10,
+    'sanitize_callback' => 'awada_sanitize_text',
+));
+Kirki::add_field('awada_theme', array(
+    'type'     => 'text',
+    'settings' => 'home_slider_shortcode',
+    'label'    => __('Plugin Shortcode', 'awada'),
+    'section'  => 'slider_sec',
+    'priority' => 10,
+    'text'     => 'awada_sanitize_text',
+));
 /* Service Options */
 Kirki::add_section('service_sec', array(
     'title'      => __('Service Options', 'awada'),
@@ -587,8 +597,7 @@ Kirki::add_field('awada_theme', array(
 	'partial_refresh' => array(
         'home_service_title' => array(
             'selector'            => '#home_service_title',
-			'container_inclusive' => true,
-            'render_callback'     => 'awada_partial_refresh_service_head'
+            'render_callback'     => function(){$awada_theme_options = awada_theme_options(); return $awada_theme_options['home_service_title'];}
         ),
     )
 ));
@@ -604,8 +613,7 @@ Kirki::add_field('awada_theme', array(
 	'partial_refresh' => array(
         'home_service_description' => array(
             'selector'            => '#home_service_description',
-			'container_inclusive' => true,
-            'render_callback'     => 'awada_partial_refresh_service_desc',
+            'render_callback'     => function(){$awada_theme_options = awada_theme_options(); return $awada_theme_options['home_service_description'];}
         ),
     )
 ));
@@ -629,7 +637,7 @@ Kirki::add_field('awada_theme', array(
     'settings'          => 'service_icon_'.$i,
     'label'             => sprintf(__('Service %s Icon','awada'), $num[$i]),
     'section'           => 'service_sec',
-    'type'              => 'text',
+    'type'              => 'fontawesome',
     'priority'          => 10,
     'transport'         => 'postMessage',
     'default'           => $awada_theme_options['service_icon_'.$i],
@@ -637,10 +645,7 @@ Kirki::add_field('awada_theme', array(
 	'partial_refresh' => array(
         'service_icon_'.$i => array(
             'selector'            => '#service_box_'.$i,
-			'container_inclusive' => true,
-            'render_callback'     => function() use($i){
-					awada_partial_refresh_service_icon($i);
-			},
+            'render_callback'     => function() use($i){$awada_theme_options = awada_theme_options(); return $awada_theme_options['service_icon_'.$i];}
         ),
     )
 	
@@ -657,10 +662,7 @@ Kirki::add_field('awada_theme', array(
 	'partial_refresh' => array(
         'service_title_'.$i => array(
             'selector'            => '#service_title_'.$i,
-			'container_inclusive' => true,
-            'render_callback'     => function() use($i){
-					awada_partial_refresh_service_title($i);
-			},
+            'render_callback'     => function() use($i){$awada_theme_options = awada_theme_options(); return $awada_theme_options['service_title_'.$i];}
         ),
     )
 	
@@ -677,10 +679,7 @@ Kirki::add_field('awada_theme', array(
 	'partial_refresh' => array(
         'service_text_'.$i => array(
             'selector'            => '#service_text_'.$i,
-			'container_inclusive' => true,
-            'render_callback'     => function() use($i){
-					awada_partial_refresh_service_text($i);
-			},
+            'render_callback'     => function() use($i){$awada_theme_options = awada_theme_options(); return $awada_theme_options['service_text_'.$i];}
         ),
     )
 	
@@ -697,10 +696,7 @@ Kirki::add_field('awada_theme', array(
 	'partial_refresh' => array(
         'service_link_'.$i => array(
             'selector'            => '#service_link_'.$i,
-			'container_inclusive' => true,
-            'render_callback'     => function() use($i){
-					awada_partial_refresh_service_link($i);
-			},
+            'render_callback'     => function() use($i){$awada_theme_options = awada_theme_options(); return $awada_theme_options['service_link_'.$i];}
         ),
     )
 	
@@ -715,13 +711,45 @@ Kirki::add_section('portfolio_sec', array(
     'capability' => 'edit_theme_options',
 ));
 Kirki::add_field('awada_theme', array(
+    'settings'          => 'home_portfolio_title',
+    'label'             => __('Portfolio Title', 'awada'),
+    'section'           => 'portfolio_sec',
+    'type'              => 'text',
+    'priority'          => 10,
+    'transport'         => 'postMessage',
+    'default'           => $awada_theme_options['home_portfolio_title'],
+    'sanitize_callback' => 'awada_sanitize_text',
+    'partial_refresh' => array(
+        'home_portfolio_title' => array(
+            'selector'            => '#portfolio_heading',
+            'render_callback'     => function(){$awada_theme_options = awada_theme_options(); return $awada_theme_options['home_portfolio_title'];}
+        ),
+    )
+));
+Kirki::add_field('awada_theme', array(
+    'settings'          => 'home_portfolio_subtitle',
+    'label'             => __('Portfolio Subtitle', 'awada'),
+    'section'           => 'portfolio_sec',
+    'type'              => 'textarea',
+    'priority'          => 10,
+    'transport'         => 'postMessage',
+    'default'           => $awada_theme_options['home_portfolio_subtitle'],
+    'sanitize_callback' => 'awada_sanitize_textarea',
+    'partial_refresh' => array(
+        'home_portfolio_subtitle' => array(
+            'selector'            => '#portfolio_sub_heading',
+            'render_callback'     => function(){$awada_theme_options = awada_theme_options(); return $awada_theme_options['home_portfolio_subtitle'];}
+        ),
+    )
+));
+Kirki::add_field('awada_theme', array(
     'settings'          => 'portfolio_post',
     'label'             => __('Select a post', 'awada'),
     'description'       => __('Select the post in which you have put shortcode.', 'awada'),
     'section'           => 'portfolio_sec',
     'type'              => 'select',
     'priority'          => 10,
-    'choices'           => awada_get_post_select(),
+    'choices'           => Kirki_Helper::get_posts(array('posts_per_page' => -1, 'orderby' => 'date', 'order' => 'DESC', 'post_type' => 'post', 'post_status' => 'publish')),
     'sanitize_callback' => 'awada_sanitize_number',
 ));
 /* Blog Options */
@@ -743,8 +771,7 @@ Kirki::add_field('awada_theme', array(
 	'partial_refresh' => array(
         'home_blog_title' => array(
             'selector'            => '#blog_heading',
-			'container_inclusive' => true,
-            'render_callback'     => 'awada_partial_refresh_blog_title'
+            'render_callback'     => function(){$awada_theme_options = awada_theme_options(); return $awada_theme_options['home_blog_title'];}
         ),
     )
 ));
@@ -760,53 +787,112 @@ Kirki::add_field('awada_theme', array(
 	'partial_refresh' => array(
         'home_blog_description' => array(
             'selector'            => '#blog_description',
-			'container_inclusive' => true,
-            'render_callback'     => 'awada_partial_refresh_blog_desc'
+            'render_callback'     => function(){$awada_theme_options = awada_theme_options(); return $awada_theme_options['home_blog_description'];}
         ),
     )
 ));
-$categories = get_categories( array(
-	'orderby' => 'name',
-	'order'   => 'ASC'
-) );
-foreach( $categories as $category ) {
-	$cats[$category->term_id] = $category->name;
-}
-$count_posts = wp_count_posts();
-$published_posts = $count_posts->publish;
+
 Kirki::add_field('awada_theme', array(
     'settings'          => 'blog_post_count',
     'label'             => __('Blog Load More Posts', 'awada'),
     'description'       => __('Show Posts On Blog Home', 'awada'),
     'help'              => __('With this option you can show blog posts according your requirement', 'awada'),
     'section'           => 'blog_sec',
-    'type'              => 'select',
+    'type'              => 'number',
     'priority'          => 10,
 	'transport'         => 'postMessage',
     'default'           => 3,
-    'choices'           => array(
-        3                => 3,
-        6                => 6,
-        9                => 9,
-        12               => 12,
-        15               => 15,
-        $published_posts => 'Show All Posts',
+    'choices'     => array(
+        'min'  => 3,
+        'max'  => 30,
+        'step' => 1,
     ),
     'sanitize_callback' => 'awada_sanitize_number',
 ));
 Kirki::add_field('awada_theme', array(
     'settings'          => 'home_post_cat',
     'label'             => __('Category', 'awada'),
-    'description'       => __('Show Posts On Blog Home According to Selected Categories', 'awada'),
-    'help'              => __('With this option you can show blog posts according your requirement', 'awada'),
+    'description'       => __('Show Posts On Home Blog According to Selected Categories', 'awada'),
+    'tooltip'              => __('With this option you can show blog posts according your requirement', 'awada'),
     'section'           => 'blog_sec',
     'type'              => 'select',
 	'transport'         => 'postMessage',
     'priority'          => 10,
-    'default'           => 0,
-	'multiple'          => 10,
-    'choices'           => $cats,
+    'default'           => array(),
+	'multiple'          => 1,
+    'choices'           => Kirki_Helper::get_terms( array('taxonomy' => 'category') ),
     'sanitize_callback' => 'awada_sanitize_number',
+));
+Kirki::add_field('awada_theme', array(
+    'settings'    => 'show_load_more_btn',
+    'label'       => __('Show Load More Button.', 'awada'),
+    'section'     => 'blog_sec',
+    'type'        => 'switch',
+    'priority'    => 10,
+    'default'     => 1,
+    'choices'     => array(
+        'on'  => __('Yes','awada'),
+        'off'  => __('No','awada'),
+    ),
+    'sanitize_callback'=>'awada_sanitize_checkbox'
+));
+Kirki::add_field('awada_theme', array(
+    'settings'    => 'home_load_post_num',
+    'label'       => __('Number of load more posts.', 'awada'),
+    'description'        => __('Number of posts you want to load on "Load More" button click', 'awada'),
+    'section'     => 'blog_sec',
+    'type'        => 'number',
+    'priority'    => 10,
+    'default'     => 3,
+    'choices'     => array(
+        'min'  => 3,
+        'max'  => 10,
+        'step' => 1,
+    ),
+    'sanitize_callback'=>'awada_sanitize_number',
+    'active_callback'=>array(
+        array(
+                'setting'  => 'show_load_more_btn',
+                'operator' => '==',
+                'value'    => true,
+            ),
+    )
+));
+
+Kirki::add_field('awada_theme', array(
+    'settings'          => 'blog_string',
+    'label'             => __('Strings Used in Home Blog section', 'awada'),
+    'section'           => 'blog_sec',
+    'type'              => 'heading',
+    'priority'          => 10,
+    'sanitize_callback' => 'awada_sanitize_text',
+));
+Kirki::add_field('awada_theme', array(
+    'settings'          => 'blog_load_more_text',
+    'label'             => __('Load More Text', 'awada'),
+    'section'           => 'blog_sec',
+    'type'              => 'text',
+    'default'           =>$awada_theme_options['blog_load_more_text'],
+    'priority'          => 10,
+    'sanitize_callback' => 'awada_sanitize_text',
+));
+Kirki::add_field('awada_theme', array(
+    'settings'          => 'blog_more_loading',
+    'label'             => __('Loading Text', 'awada'),
+    'section'           => 'blog_sec',
+    'type'              => 'text',
+    'default'           =>$awada_theme_options['blog_more_loading'],
+    'priority'          => 10,
+    'sanitize_callback' => 'awada_sanitize_text',
+));
+Kirki::add_field('awada_theme', array(
+    'settings'          => 'blog_no_more_post',
+    'label'             => __('No more post Text', 'awada'),
+    'section'           => 'blog_sec',
+    'type'              => 'text',
+    'default'           =>$awada_theme_options['blog_no_more_post'],
+    'priority'          => 10,
+    'sanitize_callback' => 'awada_sanitize_text',
 ));
 /* Footer Callout */
 Kirki::add_section('callout_sec', array(
@@ -873,8 +959,7 @@ Kirki::add_field('awada_theme', array(
 	'partial_refresh' => array(
         'home_callout_title' => array(
             'selector'            => '#callout_title',
-			'container_inclusive' => true,
-            'render_callback'     => 'awada_partial_refresh_callout_title'
+            'render_callback'     => function(){$awada_theme_options = awada_theme_options(); return $awada_theme_options['home_callout_title'];}
         ),
     )
 ));
@@ -891,8 +976,7 @@ Kirki::add_field('awada_theme', array(
 	'partial_refresh' => array(
         'home_callout_description' => array(
             'selector'            => '#callout_description',
-			'container_inclusive' => true,
-            'render_callback'     => 'awada_partial_refresh_callout_desc'
+            'render_callback'     => function(){$awada_theme_options = awada_theme_options(); return $awada_theme_options['home_callout_description'];}
         ),
     )
 ));
@@ -908,7 +992,6 @@ Kirki::add_field('awada_theme', array(
 	'partial_refresh' => array(
         'callout_btn_text' => array(
             'selector'            => '#callout_link',
-			'container_inclusive' => true,
             'render_callback'     => 'awada_partial_refresh_callout_btn'
         ),
     )
@@ -983,13 +1066,11 @@ Kirki::add_field('awada_theme', array(
 ));
 Kirki::add_field('awada_theme', array(
     'settings'          => 'social_media_header',
-    'label'             => __('Show/Hide Social Icon in Topbar', 'awada'),
-    'description'       => sprintf(__('Check out %s page to setup social menu icon in topbar', 'awada'),'<a href="'.admin_url('themes.php?page=awada-welcome').'" target="_blank">About Awada</a>'),
+    'label'             => __('Add Social Media Icons', 'awada'),
     'section'           => 'social_sec',
-    'type'              => 'switch',
+    'type'              => 'custom',
     'priority'          => 10,
-	'transport'         => 'postMessage',
-    'default'           => $awada_theme_options['social_media_header'],
+    'default'           => sprintf(__('Create a new menu and then add the URL for each social profile to your menu as a custom link. %1$s Create Now %2$s', 'awada'), '<a href="javascript:wp.customize.control( \'nav_menu_locations[social]\' ).focus();">', '</a>'),
     'sanitize_callback' => 'awada_sanitize_checkbox',
 ));
 
@@ -1279,6 +1360,29 @@ function awada_customize_register_active( $wp_customize ) {
 		'settings' => 'awada_pro',
 		'priority' => 1,
 	) ) );
+
+    class Awada_Customize_Heading extends WP_Customize_Control
+    {
+        public $type = 'heading';
+
+        public function render_content()
+        {
+            if (!empty($this->label)): ?>
+                <h3 class="awada-accordion-section-title"><?php echo esc_html($this->label); ?></h3>
+            <?php endif;
+
+            if ($this->description) {?>
+                <span class="description customize-control-description">
+                <?php echo wp_kses_post($this->description); ?>
+                </span>
+            <?php }
+        }
+    }
+    // Register our custom control with Kirki
+    add_filter('kirki/control_types', function ($controls) {
+        $controls['heading'] = 'Awada_Customize_Heading';
+        return $controls;
+    });
 }
 add_action( 'customize_register', 'awada_customize_register_active' );