+++ home-blog.php
@@ -3,58 +3,38 @@
 	<div class="container">
 		<div class="main_title">
 			<?php if ($awada_theme_options['home_blog_title'] != ""){ ?>
-			<h2 id="blog_heading"><?php echo esc_attr($awada_theme_options['home_blog_title']); ?></h2>
+			<h2 id="blog_heading"><?php echo wp_kses_post($awada_theme_options['home_blog_title']); ?></h2>
 			<hr>
 			<?php } if ($awada_theme_options['home_blog_description'] != ""){ ?>
-			<p id="blog_description" class="lead"><?php echo esc_attr($awada_theme_options['home_blog_description']); ?></p>
+			<p id="blog_description" class="lead"><?php echo wp_kses_post($awada_theme_options['home_blog_description']); ?></p>
 			<?php } ?>
 		</div><!-- end general title -->
-		<?php $imageSize = 'awada_blog_home_thumb'; ?>
-		<div id="home_blog" class="blog-masonry masonry2">
+		<div id="home_blog" class="blog-masonry ajax_posts">
 			<?php
-			$count_posts = wp_count_posts();
-			$published_posts = $count_posts->publish;
-			if(isset($awada_theme_options['home_post_cat'])){
-			$cat = $awada_theme_options['home_post_cat'];
-			}
 			$post_id = ($awada_theme_options['portfolio_post'] == ''?'':$awada_theme_options['portfolio_post']);
 			$portfolio_id = intval($post_id);
-			$args = array('post_type' => 'post', 'posts_per_page' => $published_posts, 'post__not_in' => array($portfolio_id), 'category__in'=>$cat );
+			$args = array('post_type' => 'post', 'posts_per_page' => $awada_theme_options['blog_post_count'], 'post__not_in' => array($portfolio_id));
+			if(isset($awada_theme_options['home_post_cat']) && !empty($awada_theme_options['home_post_cat'])){
+				$args['category__in']=$awada_theme_options['home_post_cat'];
+			}
             query_posts($args);
 			if (query_posts($args)) {
-			$i = 1;
-			$j = 1;
-			while (have_posts()):the_post();
-			if($i==1){
-				$pos = 'first';
-			} elseif($i==2){
-				$pos = '';
-			} else if($i==3){
-				$pos = 'last';
-				$i = 0;
-			} ?>
-			<div class="col-lg-4 <?php echo $pos; ?>" id="roww-<?php echo $j; ?>">
-				<div id="post-<?php the_ID(); ?>" <?php post_class('blog-carousel'); ?>>
-					<div class="content_entry">
-						<?php $icon = "";
-						//global $imageSize;
+			while (have_posts()):the_post();?>
+			<div class="col-lg-4 grid-item">
+				<div id="post-<?php the_ID(); ?>" <?php post_class(' blog-carousel'); ?>>
+					<div class="content_entry"><?php
 						$img_class = array('class' => 'img-responsive'); ?>
 						<?php if (has_post_thumbnail()) {
-								$icon = "fa fa-picture-o";
 								$url = wp_get_attachment_url(get_post_thumbnail_id(get_the_ID()));
-								the_post_thumbnail($imageSize, $img_class); ?>
+								the_post_thumbnail('awada_blog_home_thumb', $img_class); ?>
 							<div class="magnifier">
 								<div class="buttons">
 									<a class="st" rel="bookmark" href="<?php the_permalink(); ?>"><i class="fa fa-link"></i></a>
 									<a class="sf" data-gal="prettyPhoto[product-gallery]" href="<?php echo esc_url($url); ?>"><i class="fa fa-expand"></i></a>
 								</div><!-- end buttons -->
 							</div><!-- end magnifier -->
-							<div class="post-type">
-								<i class="<?php echo $icon; ?>"></i>
-							</div><!-- end pull-right -->
 							<?php } ?>
 					</div><!-- end content_entry -->
-					<?php if (!(class_exists('WooCommerce') && (is_cart() || is_checkout()))) {?>
 					<div class="blog-header">
 						<h3><a title="<?php the_title_attribute(); ?>" href="<?php the_permalink(); ?>"><?php the_title(); ?></a></h3>
 						<div class="blog-meta">
@@ -63,18 +43,19 @@
 							<span><i class="fa fa-user"></i> <a href="<?php echo esc_url(get_author_posts_url(get_the_author_meta('ID'))); ?>"><?php esc_attr(the_author()); ?></a></span>
 						</div><!-- end blog-meta -->
 					</div><!-- end blog-header -->
-					<?php } ?>
 					<div class="blog-desc">
-						<?php if(is_page() || is_singular()) { the_content(); } else { the_excerpt(); } ?>
+						<?php the_excerpt(); ?>
 					</div><!-- end blog-desc -->
-				</div><!-- end blog-carousel -->
-				<div class="awada_blog_shadow"></div>
-			</div><!-- end col-lg-4 -->
-			<?php $i++; $j++; endwhile;
+				</div>
+					<div class="awada_blog_shadow"></div>
+			</div><!-- end blog-carousel -->
+			<?php endwhile;
 			} ?>
 		</div><!-- end blog-masonry -->
 	</div><!-- end container -->
-	<div class="buttons padding-top text-center post-btn2">
-		<a class="btn btn-primary btn-lg append-button1 btn-shadow" title=""><?php _e('Load More Posts', 'awada'); ?></a>
-	</div>
+	<?php if($awada_theme_options['show_load_more_btn']){?>
+    <div class="buttons padding-top text-center post-btn2">
+        <button id="more_posts" type="submit" class="btn btn-primary btn-lg append-button1 btn-shadow" data-loading-text="<i class='fa fa-circle-o-notch fa-spin'></i> <?php echo esc_attr($awada_theme_options['blog_more_loading']);?>"><?php echo esc_attr($awada_theme_options['blog_load_more_text']); ?></button>
+    </div><?php 
+} ?>
 </section><!-- end white-wrapper -->